website/app/prediction.py
from flask import Blueprint, render_template, request, flash, redirect
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

bp = Blueprint('prediction', __name__)

@bp.route('/predict', methods=['POST'])
def predict():
    if request.method == 'POST':
        file = request.files['file']
        if not file:
            flash('No file selected for uploading', 'danger')
            return redirect(request.url)

        df = pd.read_csv(file)
        # Process the data and make prediction
        # ...

        # Return the prediction results
        return render_template('predict_results.html', results=results)
