website/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.2
gunicorn==22.0.0
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.24.4
packaging==24.1
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
werkzeug==3.0.3
zipp==3.19.2
google-auth 
google-auth-httplib2 
google-api-python-client
google
google-auth-oauthlib
python-dotenv